+++ src/blocks.js
@@ -148,7 +148,7 @@ CustomCommandBlockMorph, SymbolMorph, ToggleButtonMorph, DialMorph*/
 
 // Global stuff ////////////////////////////////////////////////////////
 
-modules.blocks = '2019-January-28';
+modules.blocks = '2019-February-07';
 
 var SyntaxElementMorph;
 var BlockMorph;
@@ -1966,8 +1966,10 @@ SyntaxElementMorph.prototype.fixLayout = function (silently) {
             }
         }
         if (this.width() !== initialExtent.x) {
-            affected = this.parentThatIsAnyOf(
-                [ReporterBlockMorph, CommandSlotMorph, RingCommandSlotMorph]
+            affected = this.parentThatIsA(
+                ReporterBlockMorph,
+                CommandSlotMorph,
+                RingCommandSlotMorph
             );
             if (affected) {
                 affected.fixLayout();
@@ -4692,9 +4694,25 @@ CommandBlockMorph.prototype.snap = function (hand) {
     }
 };
 
+CommandBlockMorph.prototype.prepareToBeGrabbed = function (handMorph) {
+    // +++
+    var oldPos = this.position();
+
+    nop(handMorph);
+    if (this.parent instanceof RingReporterSlotMorph) {
+        this.parent.revertToDefaultInput(this);
+        this.setPosition(oldPos);
+    }
+    CommandBlockMorph.uber.prepareToBeGrabbed.call(this, handMorph);
+};
+
 CommandBlockMorph.prototype.isStop = function () {
+    var choice;
+    if (this.selector === 'doStopThis') { // this could be cached...
+        choice = this.inputs()[0].evaluate();
+        return choice instanceof Array && choice[0].length < 12;
+    }
     return ([
-        'doStopThis',
         'doForever',
         'doReport',
         'removeClone'
@@ -8437,13 +8455,13 @@ InputSlotMorph.prototype.setContents = function (data) {
     }
     cnts.drawNew();
 
+    // remember the constant, if any
+    this.constant = isConstant ? data : null;
+
     // adjust to zebra coloring:
     if (this.isReadOnly && (this.parent instanceof BlockMorph)) {
         this.parent.fixLabelColor();
     }
-
-    // remember the constant, if any
-    this.constant = isConstant ? data : null;
 };
 
 InputSlotMorph.prototype.userSetContents = function (aStringOrFloat) {
@@ -8511,7 +8529,7 @@ InputSlotMorph.prototype.menuFromDict = function (
        			dial.action = update;
        			dial.fillColor = this.parent.color;
           		dial.setRadius(this.fontSize * 3);
-				dial.setValue(this.evaluate(), false, true);
+				dial.setValue(+this.evaluate(), false, true);
        			menu.addLine();
 			    menu.items.push(dial);
             	menu.addLine();
@@ -8914,7 +8932,11 @@ InputSlotMorph.prototype.pianoKeyboardMenu = function () {
         this.right() - (menu.width() / 2),
         this.bottom()
     ));
-    menu.selectKey(this.evaluate());
+    menu.selectKey(+this.evaluate());
+};
+
+InputSlotMorph.prototype.directionDialMenu = function () {
+    return {'§_dir': null};
 };
 
 InputSlotMorph.prototype.setChoices = function (dict, readonly) {
@@ -11586,6 +11608,8 @@ RingReporterSlotMorph.prototype.rfBorder
 RingReporterSlotMorph.prototype.edge
     = RingCommandSlotMorph.prototype.edge;
 
+RingReporterSlotMorph.prototype.enableCommandDrops = false; // +++ not finished
+
 // RingReporterSlotMorph instance creation:
 
 function RingReporterSlotMorph(isPredicate) {
@@ -11611,6 +11635,59 @@ RingReporterSlotMorph.prototype.replaceInput = function (source, target) {
         this.parent.vanishForSimilar();
     }
 };
+
+///* under construction +++
+// RingReporterSlotMorph attach targets for commands:
+
+RingReporterSlotMorph.prototype.slotAttachPoint =
+    CommandSlotMorph.prototype.slotAttachPoint;
+
+RingReporterSlotMorph.prototype.dentLeft =
+    CommandSlotMorph.prototype.dentLeft;
+
+RingReporterSlotMorph.prototype.dentCenter =
+    CommandSlotMorph.prototype.dentCenter;
+
+RingReporterSlotMorph.prototype.attachTargets = function () {
+    if (!RingReporterSlotMorph.prototype.enableCommandDrops ||
+        this.contents() instanceof ReporterBlockMorph
+    ) {
+        // don't let commands "kick out" embedded reporters
+        return [];
+    }
+    return CommandSlotMorph.prototype.attachTargets.call(this);
+};
+
+// RingReporterSlotMorph nesting for commands:
+
+RingReporterSlotMorph.prototype.nestedBlock = function (block) {
+    if (block) {
+        var nb = this.nestedBlock();
+        this.silentReplaceInput(this.children[0], block);
+        if (nb) {
+            block.bottomBlock().nextBlock(nb);
+        }
+        this.fixLayout();
+    } else {
+        return detect(
+            this.children,
+            function (child) {
+                return child instanceof BlockMorph;
+            }
+        );
+    }
+};
+
+// RingReporterSlotMorph layout:
+
+RingReporterSlotMorph.prototype.fixLayout = function () {
+    if (this.contents() instanceof CommandBlockMorph) {
+        CommandSlotMorph.prototype.fixLayout.call(this);
+    } else {
+        RingReporterSlotMorph.uber.fixLayout.call(this);
+    }
+};
+//*/
 
 // RingReporterSlotMorph drawing: